+++ docs/source/changelog.rst
@@ -7,6 +7,7 @@ Unreleased
 **Added**
 
 * `Issue #32 <https://github.com/pmacg/py-sgtl/issues/32>`_ - add tensor product method for combining graphs
+* `Issue #27 <https://github.com/pmacg/py-sgtl/issues/27>`_ - add option to plot spectrum of graph
 
 0.4.4 - 2022-01-07
 ------------------